backend/requirements.txt
Django>=4.2
djangorestframework
PyPDF2
django-cors-headers
